Moved disease_model from the docs to examples and made a symlink in the docs. The disease_model example replaces basic-infection; basic-infection is still used as a benchmark.
