+++ benches/example_basic_infection/incidence_report.rs
