+++ benches/example_basic_infection/infection_manager.rs
