+++ benches/example_basic_infection/transmission_manager.rs
