+++ docs/book/models/disease_model
@@ -1,0 +1,1 @@
+../../../examples/disease_model
