--- examples/basic-infection/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "ixa_example_basic_infection"
-version = "0.1.0"
-edition = "2021"
-
-publish = false
-
-[dependencies]
-ixa = { path = "../../../ixa" }
-ixa-derive = { path = "../../ixa-derive" }
-
-csv = "^1.3.1"
-serde = "^1.0.217"
-rand = "^0.8.5"
-rand_distr = "^0.4.3"
-paste = "^1.0.15"
-
-[[bin]]
-name = "basic_infection"
-path = "main.rs"
