+++ examples/disease_model/.gitignore
