+++ examples/disease_model/src/incidence_report.rs
