+++ examples/disease_model/src/infection_manager.rs
