+++ examples/disease_model/src/people.rs
