+++ examples/disease_model/src/transmission_manager.rs
